resouce manager returns copy of md2 files
when returning an animated object, the resource manager returns a pointer to a new copy, not the original. The copy should have the same pointer as the original to the frame data but has unique animation state data such as animation speed and current frame. This should allow sharing of animation resources between game objects and still having the ability to animate them independently.